refactor(ask): rename nclutils.questions to nclutils.ask
