+++ src/nclutils/ask/__init__.py
@@ -1,5 +1,5 @@
 """Question utilities."""
 
-from .questions import choose_multiple_from_list, choose_one_from_list
+from .ask import choose_multiple_from_list, choose_one_from_list
 
 __all__ = ["choose_multiple_from_list", "choose_one_from_list"]
